requirements.txt
torch~=1.9.0
numpy~=1.19.5
piq~=0.5.4
torchvision~=0.2.1
scipy~=1.5.3
scikit-image~=0.16.2
tqdm~=4.62.0
pandas~=1.0.3
matplotlib~=3.3.4
scikit-learn~=0.24.2
open3d~=0.13.0
termcolor~=1.1.0
seaborn~=0.11.1
easydict~=1.9
imageio~=2.9.0
plyfile~=0.7.2
torch_geometric
gitpython
piq
imageio
seaborn
ipywidgets